Use Ragel "-G2" flag by default
This massively speeds up the language lexers. It also increases the
size of the generated code, but the lexers are so simple it doesn't
really matter.

This reduces the time taken to count the latest Linux repo (at commit
bb6d3fb354c5ee8d6) from ~18 seconds to ~3.1 seconds on a quad core
i5-6400.